ingestion/whale/pyproject.toml
[project]
name = "whale"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-pubsub>=2.36.0",
    "requests>=2.33.1",
]
